+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 0.3.0 2017-03-09 [Groupin by Time]
+
+- **Added** Command to Group by second
+- **Added** Command to Group by minute
+- **Added** Command to Group by hour
+
 ## 0.2.0 2017-03-09 [Remove Date]
 
 - **Added** Command to remove Date (not time)